Improved parser for unions.
Currently:

(A | B | C[]) will often come back as C even if A or B is present due to the fact that an empty [] has a very low cost. we can attach a higher cost to throwing away objects that are not matching yourself, but do match others.

For now the solution is simply to devalue default matches as low-quality results. e.g. [] will always match worse than some poor match of A or B.

[ { ... } ] will not be devalued (only empty lists)