feat: UDF evaluator through YAML file
Implemented a UDF runner that parses a YAML file that defines functions
to be run on certain inputs, and runs Jinja expressions to calculate the
outputs for those functions when the given data matches. More
information in the sample-prices.yaml file that is also used for
testing.

Useful for computing statistics from raw usage, like how much money is
the model going to cost after a month of using it.